+++ js/rollup.config.js
@@ -1,26 +1,24 @@
-import resolve from '@rollup/plugin-node-resolve';
-import {terser} from 'rollup-plugin-terser';
-import commonjs from '@rollup/plugin-commonjs';
+import resolve from "@rollup/plugin-node-resolve";
+import terser from "@rollup/plugin-terser";
+import commonjs from "@rollup/plugin-commonjs";
+
+const outputDir = "../src/collective/bbcodesnippets/static";
 
 export default {
-	input: 'src/main.js',
-	output: [{
-		sourcemap: true,
-		format: 'iife',
-		name: 'collectivebbcodesnippets',
-		file: '../src/collective/bbcodesnippets/static/collective.bbcodesnippets.js'
-	},
+  input: "src/main.js",
+  output: [
     {
-		file: '../src/collective/bbcodesnippets/static/collective.bbcodesnippets.min.js',
-		format: 'iife',
-		name: 'version',
-		plugins: [terser()]
-	  }
-	],
-	plugins: [
-		resolve({
-			browser: true,
-		}),	
-		commonjs(),
-	]
-}
+      file: `${outputDir}/collective.bbcodesnippets.js`,
+      format: "iife",
+      name: "collectivebbcodesnippets",
+      sourcemap: true,
+    },
+    {
+      file: `${outputDir}/collective.bbcodesnippets.min.js`,
+      format: "iife",
+      name: "collectivebbcodesnippets",
+      plugins: [terser()],
+    },
+  ],
+  plugins: [resolve({ browser: true }), commonjs()],
+};